### 2018-04-12 #673 @kwwette
@arfon Thanks!

Re. potential reviewers: from the list I'd suggest perhaps @fgnievinski, @Scivision, or @jordigh for their Octave expertise. Unfortunately I don't have any off-list reviewers to suggest: in the field of gravitational wave astronomy/astrophysics we are all part of large collaborations and publish papers with 1000+ authors, so I'm a co-author with pretty much everyone who might have scientific expertise in the field.
