### 2018-04-12 #673 @kwwette
Also re. the article proof: looks good, there are a few minor issues with the formatting of citations and references, should I be trying to correct those now? I'm also not sure whether there should be a **References** header between the acknowledgements and the bibliography.
