### 2018-04-12 c511f7c
JOSS paper: minor fixes to citations and references
