+++ paper/paper.bib
@@ -8,14 +8,14 @@
 }
 
 @manual{GSL2009,
-  author =       {M. Galassi and others},
+  author =       {M. Galassi et al.},
   title =        {GNU Scientific Library Reference Manual - Third Edition},
   year =         2009,
   url =          {https://www.gnu.org/software/gsl},
 }
 
 @article{LIGOVirg2016a,
-  author =       {Abbott, B.~P. and others},
+  author =       {Abbott, B.~P. et al.},
   title =        {{Observation of Gravitational Waves from a Binary Black Hole
                   Merger}},
   journal =      {Physical Review Letters},
@@ -32,7 +32,7 @@
 }
 
 @article{LIGOVirg2017a,
-  author =       {{Abbott}, B.~P. and others},
+  author =       {{Abbott}, B.~P. et al.},
   title =        {{GW170817: Observation of Gravitational Waves from a Binary
                   Neutron Star Inspiral}},
   journal =      {Physical Review Letters},
