### 2018-04-13 #673 @kwwette
FYI I've fixed up the citations and references formatting and added a **References** header.
