### 2018-04-23 #673 @arfon
:wave: @mzevin1 would you be willing to review this submission for JOSS?
