### 2018-04-23 #673 @mzevin1
Hi @arfon, I'm a member of LIGO as well (and not too versed in continuous waves science), so I may not be best suited for review.
