### 2018-04-25 #673 @arfon
> Unfortunately I don't have any off-list reviewers to suggest: in the field of gravitational wave astronomy/astrophysics we are all part of large collaborations and publish papers with 1000+ authors, so I'm a co-author with pretty much everyone who might have scientific expertise in the field.

Thanks @kwwette. I'm curious how other journals manage to review gravitational wave related papers...?

@fgnievinski, @scivision, @jordigh - would you be willing to review this submission for JOSS? We'd only be asking you to review the software aspects of this (and not the gravitational wave physics!)
