### 2018-04-27 #707 @whedon
[REVIEW]: OctApps: a library of Octave functions for continuous gravitational-wave data analysis
**Submitting author:** @kwwette (<a href="http://orcid.org/0000-0002-4394-7179">Karl Wette</a>)
**Repository:** <a href="https://github.com/octapps/octapps" target ="_blank">https://github.com/octapps/octapps</a>
**Version:** v0.1
