### 2018-04-30 #707 @stevenrbrandt
@kwwette, I got stuck at the "make check" phase above (though I had problems with installation because of LALSuite).
```
CostFunctionsBinary.m:
CostFunctionsBinary.m:   tol = -1e-2;
CostFunctionsBinary.m:   assert ( sol_v2.Nseg, 36, tol );
CostFunctionsBinary.m:   assert ( sol_v2.Tseg, 864000, tol );
CostFunctionsBinary.m:   assert ( sol_v2.mCoh, 0.800740, tol );
CostFunctionsBinary.m:   assert ( sol_v2.mInc, 0.043971, tol );
CostFunctionsBinary.m: !!!!! test failed
CostFunctionsBinary.m: exception encountered in Fortran subroutine xgammainc_
CostFunctionsBinary.m: -------------------------------------------------
Makefile:233: recipe for target 'CostFunctionsBinary.m.test' failed
make[1]: *** [CostFunctionsBinary.m.test] Error 1
make[1]: Leaving directory '/octapps'
Makefile:186: recipe for target 'check' failed
make: *** [check] Error 1
```
I attempted to use your code from within Docker. I used Debian jessie because that was what the instructions on the LALSuite site said to do.

Here's my install script:
```
FROM debian:jessie
RUN apt-get update
RUN apt-get install -y git octave liboctave-dev swig3.0 texinfo
RUN echo deb http://www.deb-multimedia.org jessie main non-free >> /etc/apt/sources.list
RUN echo deb-src http://www.deb-multimedia.org jessie main non-free >> /etc/apt/sources.list
RUN echo deb http://httpredir.debian.org/debian/ jessie-backports main >> /etc/apt/sources.list
RUN apt-get update
RUN apt-get install --force-yes -y deb-multimedia-keyring
RUN apt-get update
RUN apt-get install -y ffmpeg libgsl-dev
RUN echo deb http://software.ligo.org/lscsoft/debian jessie contrib >> /etc/apt/sources.list
RUN echo deb-src http://software.ligo.org/lscsoft/debian jessie contrib >> /etc/apt/sources.list
RUN apt-get update
RUN apt-get install --force-yes -y lscsoft-archive-keyring
RUN apt-get update # This step missing from LAL docs, as was "force-yes" above
RUN apt-get install -y lal lal-octave
RUN git clone https://github.com/octapps/octapps.git
RUN apt-get install -y build-essential
WORKDIR /octapps
RUN make
RUN echo source /octapps/octapps-user-env.sh >> /root/.bashrc
RUN make check
```
