### 2018-05-03 octapps/octapps#12 @stevenrbrandt
Can't run tests
I left details on your JOSS review. Did you see? Thanks.
https://github.com/openjournals/joss-reviews/issues/707
