### 2018-05-04 411e00e
Update package dependencies in README.md, and update Brewfile to match
