### 2018-05-04 313af49
Makefile: use ALL_CFLAGS for list of common compile flags
