### 2018-05-04 53c1f01
Makefile: do not warn on deprecated declarations in Octave headers
