### 2018-05-04 4864f4b
Makefile: ensure compile commands fail in pkg-config fails
