### 2018-05-04 octapps/octapps#12 @kwwette
@stevenrbrandt Btw the Dockerfile you wrote is very useful - would it be okay if I added it (with some updates) to OctApps itself?
