### 2018-05-04 8faac8d
Make LALSuite scripts compatible with current jessie packages
- Affects ComputeSuperskyMetrics(), injectionRecoveryGCT(),
  and loadCandidateFile()
- Can be reverted once jessie packages are updated
