+++ src/cw-data-analysis/injectionRecoveryGCT.m
@@ -669,6 +669,11 @@ endfunction ## measureGCTmismatch()
 %!  if isempty(file_in_path(getenv("PATH"), "lalapps_HierarchSearchGCT"))
 %!    disp("skipping test: LALApps programs not available"); return;
 %!  endif
+%!  output = nthargout(2, @system, "lalapps_HierarchSearchGCT --version");
+%!  LALApps_version = versionstr2hex(nthargout(5, @regexp, output, "LALApps: ([0-9.]+)"){1}{1,1});
+%!  if LALApps_version <= 0x06210000
+%!    disp("cannot run test as version of lalapps_HierarchSearchGCT is too old"); return;
+%!  endif
 %!  oldpwd = pwd;
 %!  basedir = mkpath(tempname(tempdir));
 %!  unwind_protect
