### 2018-05-04 465b2b4
Add Dockerfile for building and testing OctApps on Debian jessie
- Thanks to Steven R. Brandt for providing the initial version of this file
