### 2018-05-05 9db733d
.travis.yml: remove debugging statement
