### 2018-05-05 4eb900e
injectionRecoveryGCT(): check for lalapps_HierarchSearchGCT in tests
