### 2018-05-05 a374d38
.travis.yml: add braces around environment variables
