+++ .travis.yml
@@ -11,7 +11,7 @@ matrix:
       compiler: clang
 
 install:
-  - if test "$TRAVIS_OS_NAME" = "linux"; then true
+  - if test "${TRAVIS_OS_NAME}" = "linux"; then true
     && sudo rm -rf /etc/apt/sources.list.d/*
     && travis_retry sudo add-apt-repository -y ppa:git-core/ppa
     && travis_retry sudo add-apt-repository -y ppa:kwwette/lalsuite-dev
@@ -23,7 +23,7 @@ install:
        liboctave-dev swig3.0 libgsl0-dev libav-tools
        texinfo
     ; fi
-  - if test "$TRAVIS_OS_NAME" = "linux"; then true
+  - if test "${TRAVIS_OS_NAME}" = "linux"; then true
     && ccache -s
     && git clone --depth=50 --branch=master https://git.ligo.org/lscsoft/lalsuite.git /tmp/lalsuite
     && pushd /tmp/lalsuite
@@ -34,7 +34,7 @@ install:
     && popd
     && source /tmp/lalsuite/_inst/etc/lalsuite-user-env.sh
     ; fi
-  - if test "$TRAVIS_OS_NAME" = "osx"; then true
+  - if test "${TRAVIS_OS_NAME}" = "osx"; then true
     && travis_retry brew update --verbose
     && travis_retry brew bundle --verbose --file=./Brewfile
     ; fi
@@ -42,20 +42,20 @@ install:
 script:
   - git log -1
   - make
-  - if test "$TRAVIS_OS_NAME" = "linux"; then true
+  - if test "${TRAVIS_OS_NAME}" = "linux"; then true
     && make check NOSKIP=1 TESTS=funcName
     && make check NOSKIP=1 TESTS=src/general/funcName.m
     && make check NOSKIP=1 TESTDIR=src/general/
     && make -j2 check NOSKIP=1
     && make -j2 html
     ; fi
-  - if test "$TRAVIS_OS_NAME" = "linux" && test "$TRAVIS_REPO_SLUG" = "octapps/octapps" && test "$TRAVIS_BRANCH" = "master" && test "$TRAVIS_EVENT_TYPE" = "push" && test "x$ACCESS_TOKEN" != x; then true
-    && echo -e "machine github.com\\nlogin kwwette\\npassword $ACCESS_TOKEN" > ~/.netrc
+  - if test "${TRAVIS_OS_NAME}" = "linux" && test "${TRAVIS_REPO_SLUG}" = "octapps/octapps" && test "${TRAVIS_BRANCH}" = "master" && test "${TRAVIS_EVENT_TYPE}" = "push" && test "x${ACCESS_TOKEN}" != x; then true
+    && echo -e "machine github.com\\nlogin kwwette\\npassword ${ACCESS_TOKEN}" > ~/.netrc
     && git clone https://github.com/octapps/octapps.github.io.git /tmp/octapps.github.io
     && rm -rf /tmp/octapps.github.io/*
     && mv -f html/* /tmp/octapps.github.io/
     && cd /tmp/octapps.github.io
-    && git config user.name "$USER"
+    && git config user.name "${USER}"
     && git config user.email "<>"
     && git add --all
     && git commit --amend -C HEAD --date="`date -R`"
