### 2018-05-05 6c95adc
.travis.yml: run Dockerfile to build and test OctApps under Debian jessie
