+++ .travis.yml
@@ -7,11 +7,17 @@ matrix:
       compiler: gcc
       sudo: required
       dist: trusty
+      env: DOCKER=no
+    - os: linux
+      compiler: gcc
+      sudo: required
+      dist: trusty
+      env: DOCKER=yes
     - os: osx
       compiler: clang
 
 install:
-  - if test "${TRAVIS_OS_NAME}" = "linux"; then true
+  - if test "${TRAVIS_OS_NAME}|${DOCKER}" = "linux|no"; then true
     && sudo rm -rf /etc/apt/sources.list.d/*
     && travis_retry sudo add-apt-repository -y ppa:git-core/ppa
     && travis_retry sudo add-apt-repository -y ppa:kwwette/lalsuite-dev
@@ -23,7 +29,7 @@ install:
        liboctave-dev swig3.0 libgsl0-dev libav-tools
        texinfo
     ; fi
-  - if test "${TRAVIS_OS_NAME}" = "linux"; then true
+  - if test "${TRAVIS_OS_NAME}|${DOCKER}" = "linux|no"; then true
     && ccache -s
     && git clone --depth=50 --branch=master https://git.ligo.org/lscsoft/lalsuite.git /tmp/lalsuite
     && pushd /tmp/lalsuite
@@ -34,6 +40,10 @@ install:
     && popd
     && source /tmp/lalsuite/_inst/etc/lalsuite-user-env.sh
     ; fi
+  - if test "${TRAVIS_OS_NAME}|${DOCKER}" = "linux|yes"; then true
+    && travis_retry sudo apt-get -qq update
+    && travis_retry sudo apt-get -qq -y -o Dpkg::Options::="--force-confnew" install docker-ce
+    ; fi
   - if test "${TRAVIS_OS_NAME}" = "osx"; then true
     && travis_retry brew update --verbose
     && travis_retry brew bundle --verbose --file=./Brewfile
@@ -41,15 +51,20 @@ install:
 
 script:
   - git log -1
-  - make
-  - if test "${TRAVIS_OS_NAME}" = "linux"; then true
+  - if test "${TRAVIS_OS_NAME}|${DOCKER}" = "linux|no"; then true
     && make check NOSKIP=1 TESTS=funcName
     && make check NOSKIP=1 TESTS=src/general/funcName.m
     && make check NOSKIP=1 TESTDIR=src/general/
     && make -j2 check NOSKIP=1
     && make -j2 html
     ; fi
-  - if test "${TRAVIS_OS_NAME}" = "linux" && test "${TRAVIS_REPO_SLUG}" = "octapps/octapps" && test "${TRAVIS_BRANCH}" = "master" && test "${TRAVIS_EVENT_TYPE}" = "push" && test "x${ACCESS_TOKEN}" != x; then true
+  - if test "${TRAVIS_OS_NAME}|${DOCKER}" = "linux|yes"; then true
+    && sudo docker build .
+    ; fi
+  - if test "${TRAVIS_OS_NAME}" = "osx"; then true
+    && make
+    ; fi
+  - if test "${TRAVIS_OS_NAME}|${DOCKER}" = "linux|no" && test "${TRAVIS_REPO_SLUG}" = "octapps/octapps" && test "${TRAVIS_BRANCH}" = "master" && test "${TRAVIS_EVENT_TYPE}" = "push" && test "x${ACCESS_TOKEN}" != x; then true
     && echo -e "machine github.com\\nlogin kwwette\\npassword ${ACCESS_TOKEN}" > ~/.netrc
     && git clone https://github.com/octapps/octapps.github.io.git /tmp/octapps.github.io
     && rm -rf /tmp/octapps.github.io/*
