### 2018-05-05 octapps/octapps#12 @kwwette
@stevenrbrandt The tests should be fixed now.

FYI the test you reported failing was because I'd forgotten to list `pkg-config` as a dependency, which meant that the build system was silently failing to build some modules. Once that was fixed, I also found some tests failed under Docker because the LALSuite packages available for Debian jessie are quite old and were missing some features. I've now made the tests backward-compatible with the Debian jessie packages.

You should now be able to run `docker build https://github.com/octapps/octapps.git` to build and test OctApps under Debian jessie. Please confirm this now works, then I'll close this issue.
