### 2018-05-08 #707 @stevenrbrandt
@kwwette I'm looking for examples and community guidelines for contributing above. I'm having trouble finding these two things. Can you help? Thanks.
