### 2018-05-09 a5458f2
Move "%!test disp()" messages onto one line
