+++ src/cw-line-veto/EstimateLinePriorsFromNormSFT.m
@@ -73,5 +73,4 @@ function [lX, freqmin, freqmax, freqbins, num_outliers, max_outlier] = EstimateL
 
 endfunction ## EstimateLinePriorsFromNormSFT()
 
-%!test
-%!  disp("no test exists for this function as it requires access to data not included in OctApps");
+%!test disp("no test exists for this function as it requires access to data not included in OctApps")
