+++ src/cw-line-veto/EstimateLinePriors.m
@@ -345,5 +345,4 @@ function params_init = check_input_parameters ( params_init )
 
 endfunction ## EstimateLinePriors()
 
-%!test
-%!  disp("no test exists for this function as it requires access to data not included in OctApps");
+%!test disp("no test exists for this function as it requires access to data not included in OctApps")
