### 2018-05-09 69f5152
__octapps_make_html__(): automatically add test code to HTML help as example code
