### 2018-05-09 3821a41
README.md: clarify existence of example usage code for OctApps functions
