### 2018-05-09 75dba59
README.md: clarifying contributing guidelines
