### 2018-05-09 b936900
fitsread(): compatibility with Octave 4.4
