### 2018-05-09 1139732
.travis.yml: fix MacOSX Octave version to 4.2.2
- SWIG does not yet support Octave 4.4
