### 2018-05-09 df8c5cd
__octapps_make_html__(): minor bug fix
