### 2018-05-09 c56ebce
__octapps_make_html__(): minor bug fixes
