+++ src/general/__octapps_make_html__.m
@@ -47,13 +47,13 @@ function __octapps_make_html__(f)
     etext = "";
     while ischar(line = fgets(fid))
       if strncmp(line, "%!assert", 8)
-        etext = strcat(etext, sprintf("\n@verbatim\n%s\n@end verbatim\n", line(9:end)));
+        etext = strcat(etext, sprintf("\n@verbatim\n%s\n@end verbatim\n", line(3:end)));
       elseif strncmp(line, "%!test disp", 11)
         continue
       elseif strncmp(line, "%!test", 6) || strncmp(line, "%!shared", 8)
         econtents = "";
         linestart = 0;
-        while ischar(line = fgets(fid)) && strncmp(line, "%!", 2)
+        while ischar(line = fgets(fid)) && strncmp(line, "%!", 2) && !(strncmp(line, "%!test", 6) || strncmp(line, "%!shared", 8))
           line = line(3:end);
           if linestart == 0 && any(!isspace(line))
             linestart = min(find(!isspace(line)));
@@ -66,14 +66,16 @@ function __octapps_make_html__(f)
         etext = strcat(etext, sprintf("\n@verbatim\n%s\n@end verbatim\n", strtrim(econtents)));
       endif
     endwhile
-    if length(strfind(etext, "@verbatim"))
-      etext = strcat("@heading Examples\n", etext);
-    else
-      etext = strcat("@heading Example\n", etext);
+    if length(etext) > 0
+      if length(strfind(etext, "@verbatim"))
+        etext = strcat("@heading Examples\n", etext);
+      else
+        etext = strcat("@heading Example\n", etext);
+      endif
+      endhtext = strfind(htext, "@end deftypefn");
+      assert(length(endhtext) == 1);
+      htext = cstrcat(htext(1:endhtext-1), etext, "\n", htext(endhtext:end));
     endif
-    endhtext = strfind(htext, "@end deftypefn");
-    assert(length(endhtext) == 1);
-    htext = cstrcat(htext(1:endhtext-1), etext, "\n", htext(endhtext:end));
   endif
   of = fullfile(getenv("OCTAPPS_TMPDIR"), sprintf("%s.texi", f));
   fid = fopen(of, "w");
