### 2018-05-09 7e6276c
GetNormSFTPowerFiles(): fix documentation
