### 2018-05-09 ad6e435
fitsread(): fix documentation
