+++ src/file-handling/fitsread.cc
@@ -37,9 +37,11 @@
 extern "C" int fffree(void*, int *);
 
 static const char *const fitsread_usage = "-*- texinfo -*- \n\
-@deftypefn {Loadable Function} {@var{data} =} fitsread(@var{filename})\n\
+@deftypefn {Loadable Function} {@var{data} =} fitsread ( @var{filename} )\n\
 \n\
 Load data from a FITS (Flexible Image Transport System) file.\n\
+\n\
+@heading Examples\n\
 \n\
 @example\n\
 data = fitsread(\"results.fits\");             # Load all data in \"results.fits\"\n\
