### 2018-05-09 24b0b69
SensitivityDepthBayesian: new sensitivity estimation function for Bayesian ULs
- also added a switch to SensitivityDepthStackSlide for Bayesian case
