+++ src/cw-sensitivity/SensitivityDepthBayesian.m
@@ -1,0 +1,232 @@
+## Copyright (C) 2016, 2017 Christoph Dreissigacker
+## Copyright (C) 2011, 2016 Karl Wette
+##
+## This program is free software; you can redistribute it and/or modify
+## it under the terms of the GNU General Public License as published by
+## the Free Software Foundation; either version 2 of the License, or
+## (at your option) any later version.
+##
+## This program is distributed in the hope that it will be useful,
+## but WITHOUT ANY WARRANTY; without even the implied warranty of
+## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+## GNU General Public License for more details.
+##
+## You should have received a copy of the GNU General Public License
+## along with with program; see the file COPYING. If not, write to the
+## Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
+## MA  02111-1307  USA
+
+## Calculate sensitivity in terms of the sensitivity depth.
+## Syntax:
+##   Depth = SensitivityDepthBayesian("opt", val, ...)
+## where:
+##   Depth    = SensitivityDepth
+##   pd_Depth = calculated false dismissal probability
+## and where options are:
+##   "pd"     = false dismissal probability
+##   "Ns"     = number of segments
+##   "Tdata"  = total amount of data used in seconds
+##   "Rsqr"   = histogram of SNR "geometric factor" R^2,
+##              computed using SqrSNRGeometricFactorHist(),
+##              or scalar giving mean value of R^2
+##   "stat"   = detection statistic, one of:
+##              * {"ChiSqr", "opt", val, ...}
+##                  chi^2 statistic, e.g. the F-statistic, see
+##                  SensitivityChiSqrFDP() for options
+##              * {"HoughFstat", "opt", val, ...}
+##                  Hough on the F-statistic, see
+##                  SensitivityHoughFstatFDP() for options
+##   "prog"   = show progress updates
+##  "misHist" = mismatch histograms (default: no mismatch)
+function Depth = SensitivityDepthBayesian(varargin)
+
+  ## parse options
+  parseOptions(varargin,
+	       {"pd", "real,strictunit,column"},
+	       {"Ns", "integer,strictpos,matrix"},
+	       {"Tdata","real, matrix"},
+	       {"Rsqr", "a:Hist", []},
+	       {"misHist","acell:Hist", []},
+	       {"stat", "cell,vector"},
+	       {"prog", "logical,scalar", false},
+	       []);
+  assert(histDim(Rsqr) == 1, "%s: R^2 must be a 1D histogram", funcName);                 #add for mismatch
+  assert(length(stat) > 1 && ischar(stat{1}), "%s: first element of 'stat' must be a string", funcName);
+  assert(isempty(misHist) || size(Ns,2) == length(misHist),"#stages unclear, #columns in Nseg must match #mismatch histograms.\n");
+
+  ## detect number of stages
+  stages = size(Ns,2);
+
+  ## select a detection statistic, we don't want to change the vector pd because here the columns are still the different stages
[…]
