### 2018-05-09 a498ea0
Replace @verbatim with @example for Texinfo help examples
