### 2018-05-09 e09730e
Remove duplicate @headings in Texinfo help
