### 2018-05-09 94940cb
SuperskyMetricsCache(): fix documentation
