### 2018-05-09 #707 @kwwette
@stevenrbrandt Examples: each Octave function has embedded tests which give some ideas as to usage. I've now updated the HTML documentation at https://octapps.github.io/ to include those examples as part of the HTML documentation. See e.g. https://octapps.github.io/cw_002doptimal_002dsearch_002dsetup.html#cw_002doptimal_002dsearch_002dsetup.

There are brief guidelines for contributing in the README.md, pointing to the OctApps issue tracker on GitHub for posting issues/bugs and submitted patches for fixes/new features.
