### 2018-05-09 e9db2a0
Further minor fixes to HTML documentation generation
