+++ src/cw-sensitivity/SensitivityDepthStackSlide.m
@@ -47,6 +47,19 @@
 ## (Note: @var{Tdata} = @var{Nsft} * @var{Tsft}, where @var{Nsft} is the total number of
 ## SFTs of length @var{Tsft} used in the search, from all @var{detectors})
 ##
+## @example
+## ## two different setups with 5 stages (5 columns, 2 rows)
+## Nseg = [90,90,44,44,22;100,100,50,50,25]
+## ## different for every stage but the same for every trial
+## Tdata = [ NSFT*1800, NSFT*900, NSFT*3600, NSFT*1800, NSFT*1800]
+## ## as we have 5 stages there must be five thrqesholds
+## avg2Fth = [6.109,6.109,7.38,8.82,15]
+## ## we also need one mismatch histogram per stage
+## misHist = @{mismatch1, mismatch2, mismatch3, mismatch4, mismatch5@}
+## ## a column with two false dimissal probabilitites, one for each trial
+## pFD = [0.1,0.05]'
+## @end example
+##
 ## @item misHist
 ## cell array of mismatch histograms, one for every stage, produced using @command{Hist()}
 ##
@@ -70,21 +83,6 @@
 ## source declination (default: all-sky)
 ##
 ## @end table
-##
-## @heading Example
-##
-## @example
-## ## two different setups with 5 stages (5 columns, 2 rows)
-## Nseg = [90,90,44,44,22;100,100,50,50,25]
-## ## different for every stage but the same for every trial
-## Tdata = [ NSFT*1800, NSFT*900, NSFT*3600, NSFT*1800, NSFT*1800]
-## ## as we have 5 stages there must be five thrqesholds
-## avg2Fth = [6.109,6.109,7.38,8.82,15]
-## ## we also need one mismatch histogram per stage
-## misHist = @{mismatch1, mismatch2, mismatch3, mismatch4, mismatch5@}
-## ## a column with two false dimissal probabilitites, one for each trial
-## pFD = [0.1,0.05]'
-## @end example
 ##
 ## @end deftypefn
 
