### 2018-05-14 #707 @stevenrbrandt
The current Dockerfile builds but does not run
```[sbrandt@wothw octapps]$ docker run -it octapps bash
bash: /octapps/octapps-user-env.sh: No such file or directory
root@e9b061c40f66:/tmp/octapps# find / -name octapps-user-env.sh
/tmp/octapps/octapps-user-env.sh
```
That should be easy to fix.
