### 2018-05-14 #707 @stevenrbrandt
While the reference guide and examples seem reasonable, I would like to see a higher level introduction to how to use the software. I don't think a tutorial is required by JOSS, but it would be nice to have. This is the sort of thing I would need to check off the "Functionality documentation" check box.
