### 2018-05-14 #707 @kwwette
@stevenrbrandt : Re. Dockerfile - thanks for spotting that, this should be fixed in the repository shortly.

Re. examples - I'll have a look at adding some more explanatory text to at least some of the examples, e.g. the ones that are gravitational-wave research focused, like the ones you referenced. I'm not sure I'll be able to do this for all the functions in OctApps though.

Re. the tutorial - I'll have to have a think about this. Within its specific scientific research area, OctApps is quite a general package which can be used to answer different scientific questions, so i.e. there's not a singular usage case which would be appropriate for a tutorial. The examples in many of the codes are meant to show how to answer different questions in gravitational-wave research; one of these could be expanded into a more complete worked example. I'm not sure I could do this for all the research-focused functions in OctApps however.

/cc @ReinhardPrix (co-author on this paper) - any ideas for a problem / material we could use for an OctApps tutorial / worked example?
