### 2018-05-15 #707 @dbkeitel
@kwwette @ReinhardPrix : Sensitivity depth estimation would be a nice case for a tutorial, as it could be useful for a wider user group at least within Continuous Waves.

Though I guess with "a higher level introduction to how to use the software" @stevenrbrandt might have rather meant things like links to general octave documentation/tutorials, a description of often-used data structures within our package (if any, iirc only the histogram class is shared widely between functions?), etc...? Or at least I'd personally think that would be the most appropriate kind of high-level docs for a package that's more a collection of small tools than a single tool for a specific purpose.
