### 2018-05-25 d9acdca
Makefile: use fixed /tmp directories for quicker check/html targets when NOCLEANUP=1
