### 2018-05-25 ca9ce56
Rename OctApps HTML documentation to "Reference Manual"
