### 2018-06-03 eaf067d
Tutorials on parseOptions() and Hist
