### 2018-06-03 d11fce1
Minor fixes to doc/tutorials.texi, README.md, and JOSS paper
