### 2018-06-03 #707 @kwwette
@stevenrbrandt @danielskatz Apologies for the delay. I have now added a tutorials section to the OctApps reference manual: see https://octapps.github.io/Tutorials.html. There are tutorials on two general-purpose features of OctApps, and several tutorials on the more research-specific functions which illustrate OctApps's usage in answering research questions in continuous gravitational-wave research.

For more general Octave tutorials, I'd recommend starting with the Octave manual at https://octave.org/doc/v4.4.0/.

I hope this is enough additional material to complete the OctApps review. Please let me know if you have any further questions.
