### 2018-06-03 #707 @kwwette
@danielskatz @arfon I have a question regarding JOSS copyright policy: does JOSS permit the final PDF version of journal articles to be cross-posted to arXiv (https://arxiv.org/)? Thanks!
