### 2018-06-03 #707 @danielskatz
@kwwette - yes, the paper is CC-BY licensed by you, so you can do whatever else you want with it.
