### 2018-06-04 #707 @stevenrbrandt
OK, I think things look good now. @danielskatz shall I close this issue?
