### 2018-06-05 #707 @kwwette
@danielskatz - I have now made the minor changes to the paper, and am happy to move forward with the publication process. I have also made a v0.2 release of OctApps, in light of the changes that have been made during this review - if possible it would be good to use that version number (v0.2) for the JOSS paper.

@stevenrbrandt - Back to you to confirm your sign-off. Thanks again!
