### 2018-06-06 #707 @kwwette
@danielskatz - I've also created an archive on Zenodo for version 0.2, the DOI is 10.5281/zenodo.1283525
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1283525.svg)](https://doi.org/10.5281/zenodo.1283525)
