### 2018-06-06 39fde80
Makefile and LaTeX version of paper (used for arXiv submission)
- Adapted from Matt's psrqpy package:
  https://github.com/mattpitkin/psrqpy/commit/23d818a1f3a909fb22b0af274f852a818525e27e
