### 2018-06-06 #707 @kwwette
@arfon @danielskatz @stevenrbrandt - Excellent, thank you!
